+++ bonus/srcs/raycast/ray_cylinder_hit.c
@@ -6,7 +6,7 @@
 /*   By: fmauguin <fmauguin@student.42.fr >         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/29 17:39:27 by fmauguin          #+#    #+#             */
-/*   Updated: 2022/09/14 16:35:16 by fmauguin         ###   ########.fr       */
+/*   Updated: 2022/09/15 18:19:22 by fmauguin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,20 +16,20 @@
 #include "libft.h"
 #include <math.h>
 
-static float	set_t(t_pos *e, t_pos cy, t_ray *ray, t_pos cy_vec3)
+static double	set_t(t_pos *e, t_pos cy, t_ray *ray, t_pos cy_vec3)
 {
 	vector_ab(ray->origin, cy, e);
 	vector_equal(cy, e + 1);
 	return (dot_product(e[0], cy_vec3));
 }
 
-static float	set_t2(float dot, float denom, t_bool *dir, t_bool sign)
+static double	set_t2(double dot, double denom, t_bool *dir, t_bool sign)
 {
 	*dir = sign;
 	return (dot / denom);
 }
 
-static void	set_extremity_hitpos(t_ray *ray, t_hit *hit, float t)
+static void	set_extremity_hitpos(t_ray *ray, t_hit *hit, double t)
 {
 	vector_equal(ray->dir, &hit->pos);
 	vector_scale(t, &hit->pos);
@@ -58,14 +58,14 @@ static void	set_extremity_hit(t_vol *cy, t_hit *hit, t_bool dir, t_ray *ray)
 
 void	check_cylinder_extremity(t_vol *cy, t_pos cy_top, t_ray *ray)
 {
-	float	denom;
+	double	denom;
 	t_bool	dir;
-	float	t;
+	double	t;
 	t_pos	e[2];
 	t_hit	hit;
 
 	denom = dot_product(cy->vec3, ray->dir);
-	if (fabsf(denom) > 0.0001f)
+	if (fabs(denom) > 0.0001f)
 	{
 		t = set_t2(set_t(e, cy->pos, ray, cy->vec3), denom, &dir, true);
 		if (t < 0.00001f || (t >= 0.00001f
