+++ bonus/srcs/utils/bbox_utils.c
@@ -6,7 +6,7 @@
 /*   By: fmauguin <fmauguin@student.42.fr >         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/26 12:23:53 by fmauguin          #+#    #+#             */
-/*   Updated: 2022/08/26 22:01:53 by fmauguin         ###   ########.fr       */
+/*   Updated: 2022/09/15 18:10:37 by fmauguin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ void	bbox_center(const t_box *box, t_pos *center)
 	center->z = box->min.z + center->z / 2;
 }
 
-float	bbox_expend(const t_box *box)
+double	bbox_expend(const t_box *box)
 {
 	return (dist_ab(&box->min, &box->max));
 }
