+++ bonus/srcs/volume/bounds.c
@@ -6,7 +6,7 @@
 /*   By: fmauguin <fmauguin@student.42.fr >         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/24 14:54:16 by amahla            #+#    #+#             */
-/*   Updated: 2022/09/13 13:49:09 by amahla           ###   ########.fr       */
+/*   Updated: 2022/09/15 18:10:37 by fmauguin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 
 void	sphere_bounds(t_vol *sp)
 {
-	float	r;
+	double	r;
 
 	r = sp->d / 2;
 	set_vector(-r, -r, -r, &sp->box.min);
